requirements.txt
numpy>=1.24.0,<2.0.0
opencv-python>=4.5.0,<5.0.0
opencv-contrib-python>=4.5.0
opencv-python-headless>=4.5.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
Pillow>=9.5.0,<10.0.0
scikit-learn>=1.5.0,<2.0.0
python-multipart>=0.0.19
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0